remove blog link
we removed blog link from the contact page as 1) this is not really contact info and 2) this information is now on the main page of the site. It would have been redundant to leave it there. We also put the medium link at the same level as the quora link.